First upload of hlsvdpropy package - godspeed
This is a 'pure Python' implementation of the algorithm provided by the HLSVDPRO version 2.x package.  The HLSVDPROPY package computes a 'sum of lorentzians' model for the complex 'signals' data passed in via a 'black box' state space approach. We are using the scipy.linalg SVD libraries, rather than the PROPACK Fortran libraries used in the HLSVDPRO package, but the algorithm is otherwise similar.